third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/workers/shared-worker-response-type-blob-expected.txt

Request a file as the Blob type with asynchronous XMLHttpRequest in a shared worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/xmlhttprequest-response-type-blob.js
PASS [Worker] req.status is 200
PASS [Worker] req.response.size is 4
PASS [Worker] buffer is "PASS"
PASS [Worker] XHR with responseType="blob" succeeded
PASS successfullyParsed is true

TEST COMPLETE
